Refactor doc favourites to explorer favourites. Closes #3242
