+++ stroom-app/src/main/java/stroom/app/guice/CoreModule.java
@@ -1,6 +1,7 @@
 package stroom.app.guice;
 
-import stroom.docstore.fav.impl.db.DocFavDbModule;
+import stroom.explorer.fav.impl.ExplorerFavModule;
+import stroom.explorer.fav.impl.db.ExplorerFavDbModule;
 
 import com.google.inject.AbstractModule;
 
@@ -43,8 +44,8 @@ public class CoreModule extends AbstractModule {
         install(new stroom.dictionary.impl.DictionaryModule());
         install(new stroom.docstore.impl.DocStoreModule());
         install(new stroom.docstore.impl.db.DocStoreDbPersistenceModule());
-        install(new stroom.docstore.fav.impl.DocFavModule());
-        install(new stroom.docstore.fav.impl.db.DocFavDbModule());
+        install(new ExplorerFavModule());
+        install(new ExplorerFavDbModule());
         install(new stroom.event.logging.impl.EventLoggingModule());
         install(new stroom.event.logging.rs.impl.RestResourceAutoLoggerModule());
         install(new stroom.explorer.impl.ExplorerModule());
