--- stroom-docstore/stroom-docstore-fav-api/build.gradle
@@ -1,13 +1,0 @@
-ext.moduleName = "stroom.docstore.fav.api"
-
-dependencies {
-    implementation project(':stroom-core-shared')
-    implementation project(':stroom-docref')
-    implementation project(':stroom-util-shared')
-
-    implementation libs.jackson_annotations
-    implementation libs.jaxb_api
-
-    testRuntimeOnly libs.jakarta_activation
-    testRuntimeOnly libs.jaxb_impl
-}
