--- stroom-docstore/stroom-docstore-fav-impl-db-jooq/src/main/java/stroom/docstore/fav/impl/db/jooq/Tables.java
@@ -1,20 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package stroom.docstore.fav.impl.db.jooq;
-
-
-import stroom.docstore.fav.impl.db.jooq.tables.DocFavourite;
-
-
-/**
- * Convenience access to all tables in stroom.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Tables {
-
-    /**
-     * The table <code>stroom.doc_favourite</code>.
-     */
-    public static final DocFavourite DOC_FAVOURITE = DocFavourite.DOC_FAVOURITE;
-}
