--- stroom-docstore/stroom-docstore-fav-impl-db-jooq/src/main/java/stroom/docstore/fav/impl/db/jooq/tables/records/DocFavouriteRecord.java
@@ -1,255 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package stroom.docstore.fav.impl.db.jooq.tables.records;
-
-
-import org.jooq.Field;
-import org.jooq.Record1;
-import org.jooq.Record5;
-import org.jooq.Row5;
-import org.jooq.impl.UpdatableRecordImpl;
-
-import stroom.docstore.fav.impl.db.jooq.tables.DocFavourite;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class DocFavouriteRecord extends UpdatableRecordImpl<DocFavouriteRecord> implements Record5<Long, String, String, String, Long> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * Setter for <code>stroom.doc_favourite.id</code>.
-     */
-    public void setId(Long value) {
-        set(0, value);
-    }
-
-    /**
-     * Getter for <code>stroom.doc_favourite.id</code>.
-     */
-    public Long getId() {
-        return (Long) get(0);
-    }
-
-    /**
-     * Setter for <code>stroom.doc_favourite.doc_type</code>.
-     */
-    public void setDocType(String value) {
-        set(1, value);
-    }
-
-    /**
-     * Getter for <code>stroom.doc_favourite.doc_type</code>.
-     */
-    public String getDocType() {
-        return (String) get(1);
-    }
-
-    /**
-     * Setter for <code>stroom.doc_favourite.doc_uuid</code>.
-     */
-    public void setDocUuid(String value) {
-        set(2, value);
-    }
-
-    /**
-     * Getter for <code>stroom.doc_favourite.doc_uuid</code>.
-     */
-    public String getDocUuid() {
-        return (String) get(2);
-    }
-
-    /**
-     * Setter for <code>stroom.doc_favourite.user_uuid</code>.
-     */
-    public void setUserUuid(String value) {
-        set(3, value);
-    }
-
-    /**
-     * Getter for <code>stroom.doc_favourite.user_uuid</code>.
-     */
-    public String getUserUuid() {
-        return (String) get(3);
-    }
-
-    /**
-     * Setter for <code>stroom.doc_favourite.create_time_ms</code>.
-     */
-    public void setCreateTimeMs(Long value) {
-        set(4, value);
-    }
-
-    /**
-     * Getter for <code>stroom.doc_favourite.create_time_ms</code>.
-     */
-    public Long getCreateTimeMs() {
-        return (Long) get(4);
-    }
-
-    // -------------------------------------------------------------------------
-    // Primary key information
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Record1<Long> key() {
-        return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record5 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row5<Long, String, String, String, Long> fieldsRow() {
-        return (Row5) super.fieldsRow();
-    }
-
-    @Override
-    public Row5<Long, String, String, String, Long> valuesRow() {
-        return (Row5) super.valuesRow();
-    }
-
-    @Override
-    public Field<Long> field1() {
-        return DocFavourite.DOC_FAVOURITE.ID;
-    }
-
-    @Override
-    public Field<String> field2() {
-        return DocFavourite.DOC_FAVOURITE.DOC_TYPE;
-    }
-
-    @Override
-    public Field<String> field3() {
-        return DocFavourite.DOC_FAVOURITE.DOC_UUID;
-    }
-
-    @Override
-    public Field<String> field4() {
-        return DocFavourite.DOC_FAVOURITE.USER_UUID;
-    }
-
-    @Override
-    public Field<Long> field5() {
-        return DocFavourite.DOC_FAVOURITE.CREATE_TIME_MS;
-    }
-
-    @Override
-    public Long component1() {
-        return getId();
-    }
-
-    @Override
-    public String component2() {
-        return getDocType();
-    }
-
-    @Override
-    public String component3() {
-        return getDocUuid();
-    }
-
-    @Override
-    public String component4() {
-        return getUserUuid();
-    }
-
-    @Override
-    public Long component5() {
-        return getCreateTimeMs();
-    }
-
-    @Override
-    public Long value1() {
-        return getId();
-    }
-
-    @Override
-    public String value2() {
-        return getDocType();
-    }
-
-    @Override
-    public String value3() {
-        return getDocUuid();
-    }
-
-    @Override
-    public String value4() {
-        return getUserUuid();
-    }
-
-    @Override
-    public Long value5() {
-        return getCreateTimeMs();
-    }
-
-    @Override
-    public DocFavouriteRecord value1(Long value) {
-        setId(value);
-        return this;
-    }
-
-    @Override
-    public DocFavouriteRecord value2(String value) {
-        setDocType(value);
-        return this;
-    }
-
-    @Override
-    public DocFavouriteRecord value3(String value) {
-        setDocUuid(value);
-        return this;
-    }
-
-    @Override
-    public DocFavouriteRecord value4(String value) {
-        setUserUuid(value);
-        return this;
-    }
-
-    @Override
-    public DocFavouriteRecord value5(Long value) {
-        setCreateTimeMs(value);
-        return this;
-    }
-
-    @Override
-    public DocFavouriteRecord values(Long value1, String value2, String value3, String value4, Long value5) {
-        value1(value1);
-        value2(value2);
-        value3(value3);
-        value4(value4);
-        value5(value5);
-        return this;
-    }
-
-    // -------------------------------------------------------------------------
-    // Constructors
-    // -------------------------------------------------------------------------
-
-    /**
-     * Create a detached DocFavouriteRecord
-     */
-    public DocFavouriteRecord() {
-        super(DocFavourite.DOC_FAVOURITE);
-    }
-
-    /**
-     * Create a detached, initialised DocFavouriteRecord
-     */
-    public DocFavouriteRecord(Long id, String docType, String docUuid, String userUuid, Long createTimeMs) {
-        super(DocFavourite.DOC_FAVOURITE);
-
-        setId(id);
-        setDocType(docType);
-        setDocUuid(docUuid);
-        setUserUuid(userUuid);
-        setCreateTimeMs(createTimeMs);
-    }
-}
