--- stroom-docstore/stroom-docstore-fav-impl-db/src/main/java/stroom/docstore/fav/impl/db/DocFavDbModule.java
@@ -1,15 +1,0 @@
-package stroom.docstore.fav.impl.db;
-
-import stroom.docstore.fav.impl.DocFavDao;
-
-import com.google.inject.AbstractModule;
-
-public class DocFavDbModule extends AbstractModule {
-
-    @Override
-    protected void configure() {
-        super.configure();
-
-        bind(DocFavDao.class).to(DocFavDaoImpl.class);
-    }
-}
