--- stroom-docstore/stroom-docstore-fav-impl/build.gradle
@@ -1,46 +1,0 @@
-ext.moduleName = 'stroom.docstore.fav.impl'
-
-dependencies {
-    implementation project(':stroom-core-shared')
-    implementation project(':stroom-docref')
-    implementation project(':stroom-docstore:stroom-docstore-fav-api')
-    implementation project(':stroom-docstore:stroom-docstore-fav-impl-db-jooq')
-    implementation project(':stroom-explorer:stroom-docrefinfo-api')
-    implementation project(':stroom-explorer:stroom-explorer-api')
-    implementation project(':stroom-event-logging:stroom-event-logging-api')
-    implementation project(':stroom-event-logging-rs:stroom-event-logging-rs-api')
-    implementation project(':stroom-security:stroom-security-api')
-    implementation project(':stroom-util')
-    implementation project(':stroom-util-shared')
-
-    implementation libs.eventLogging
-    implementation libs.guice5
-    implementation libs.jackson_annotations
-    implementation libs.jakarta_servlet_api
-    implementation libs.javax_inject
-    implementation libs.jersey_common
-    implementation libs.jaxb_api
-    implementation libs.restygwt
-    implementation libs.slf4j_api
-    implementation libs.swagger_annotations
-    implementation libs.ws_rs_api
-
-    testImplementation project(':stroom-security:stroom-security-mock')
-
-    testImplementation libs.assertj_core
-    testImplementation libs.junit_jupiter_api
-    testImplementation libs.mockito_core
-    testImplementation libs.mockito_junit_jupiter
-
-    runtimeOnly libs.mysql_connector_java
-
-    // The following logging libs are needed when running junits outside dropwizard
-    testRuntimeOnly libs.jakarta_activation
-    testRuntimeOnly libs.jaxb_impl
-    testRuntimeOnly libs.jcl_over_slf4j
-    testRuntimeOnly libs.jul_to_slf4j
-    testRuntimeOnly libs.junit_jupiter_engine
-    testRuntimeOnly libs.log4j_over_slf4j
-    testRuntimeOnly libs.logback_classic
-    testRuntimeOnly libs.logback_core
-}
