--- stroom-docstore/stroom-docstore-fav-impl/src/main/java/stroom/docstore/fav/impl/DocFavModule.java
@@ -1,19 +1,0 @@
-package stroom.docstore.fav.impl;
-
-import stroom.docstore.fav.api.DocFavService;
-import stroom.util.guice.RestResourcesBinder;
-
-import com.google.inject.AbstractModule;
-
-public class DocFavModule extends AbstractModule {
-
-    @Override
-    protected void configure() {
-        super.configure();
-
-        bind(DocFavService.class).to(DocFavServiceImpl.class);
-
-        RestResourcesBinder.create(binder())
-                .bind(DocFavResourceImpl.class);
-    }
-}
