--- stroom-docstore/stroom-docstore-impl-db/src/main/resources/stroom/docstore/impl/db/migration/V07_01_00_002__Doc_favourite.sql
@@ -1,43 +1,0 @@
--- ------------------------------------------------------------------------
--- Copyright 2020 Crown Copyright
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
--- ------------------------------------------------------------------------
-
--- stop note level warnings about objects (not)? existing
-set @old_sql_notes = @@sql_notes, sql_notes = 0;
-
-CREATE TABLE IF NOT EXISTS `doc_favourite`
-(
-    `id`             bigint       NOT NULL AUTO_INCREMENT,
-    `doc_type`       varchar(255) NOT NULL,
-    `doc_uuid`       varchar(255) NOT NULL,
-    `user_uuid`      varchar(255) NOT NULL,
-    `create_time_ms` bigint       NOT NULL,
-    PRIMARY KEY (`id`),
-    KEY `doc_favourite_fk_doc_type_doc_uuid` (`doc_type`, `doc_uuid`),
-    KEY `doc_favourite_fk_user_uuid` (`user_uuid`),
-    UNIQUE KEY `doc_favourite_fk_explorer_node_type_explorer_node_uuid_user_uuid` (`doc_type`, `doc_uuid`, `user_uuid`),
-    CONSTRAINT `doc_favourite_fk_user_uuid` FOREIGN KEY (`user_uuid`)
-        REFERENCES `stroom`.`stroom_user` (`uuid`)
-        ON DELETE CASCADE,
-    CONSTRAINT `doc_favourite_fk_explorer_node_type_explorer_node_uuid` FOREIGN KEY (`doc_type`, `doc_uuid`)
-        REFERENCES `stroom`.`explorer_node` (`type`, `uuid`)
-        ON DELETE CASCADE
-) ENGINE = InnoDB
-  AUTO_INCREMENT = 446
-  DEFAULT CHARACTER SET utf8mb4
-  COLLATE utf8mb4_0900_ai_ci;
-
--- Reset to the original value
-SET SQL_NOTES = @OLD_SQL_NOTES;
