+++ stroom-explorer/stroom-explorer-fav-api/build.gradle
@@ -1,0 +1,13 @@
+ext.moduleName = "stroom.explorer.fav.api"
+
+dependencies {
+    implementation project(':stroom-core-shared')
+    implementation project(':stroom-docref')
+    implementation project(':stroom-util-shared')
+
+    implementation libs.jackson_annotations
+    implementation libs.jaxb_api
+
+    testRuntimeOnly libs.jakarta_activation
+    testRuntimeOnly libs.jaxb_impl
+}
