+++ stroom-explorer/stroom-explorer-fav-api/src/main/java/stroom/explorer/fav/api/ExplorerFavService.java
@@ -1,0 +1,16 @@
+package stroom.explorer.fav.api;
+
+import stroom.docref.DocRef;
+
+import java.util.List;
+
+public interface ExplorerFavService {
+
+    void create(final DocRef docRef);
+
+    void delete(final DocRef docRef);
+
+    List<DocRef> getUserFavourites();
+
+    boolean isFavourite(final DocRef docRef);
+}
