+++ stroom-explorer/stroom-explorer-fav-impl-db-jooq/src/main/java/stroom/explorer/fav/impl/db/jooq/Keys.java
@@ -1,0 +1,29 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package stroom.explorer.fav.impl.db.jooq;
+
+
+import org.jooq.TableField;
+import org.jooq.UniqueKey;
+import org.jooq.impl.DSL;
+import org.jooq.impl.Internal;
+
+import stroom.explorer.fav.impl.db.jooq.tables.ExplorerFavourite;
+import stroom.explorer.fav.impl.db.jooq.tables.records.ExplorerFavouriteRecord;
+
+
+/**
+ * A class modelling foreign key relationships and constraints of tables in
+ * stroom.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class Keys {
+
+    // -------------------------------------------------------------------------
+    // UNIQUE and PRIMARY KEY definitions
+    // -------------------------------------------------------------------------
+
+    public static final UniqueKey<ExplorerFavouriteRecord> KEY_EXPLORER_FAVOURITE_EXPLORER_FAVOURITE_FK_EN_TYPE_EN_UUID_USER_UUID = Internal.createUniqueKey(ExplorerFavourite.EXPLORER_FAVOURITE, DSL.name("KEY_explorer_favourite_explorer_favourite_fk_en_type_en_uuid_user_uuid"), new TableField[] { ExplorerFavourite.EXPLORER_FAVOURITE.DOC_TYPE, ExplorerFavourite.EXPLORER_FAVOURITE.DOC_UUID, ExplorerFavourite.EXPLORER_FAVOURITE.USER_UUID }, true);
+    public static final UniqueKey<ExplorerFavouriteRecord> KEY_EXPLORER_FAVOURITE_PRIMARY = Internal.createUniqueKey(ExplorerFavourite.EXPLORER_FAVOURITE, DSL.name("KEY_explorer_favourite_PRIMARY"), new TableField[] { ExplorerFavourite.EXPLORER_FAVOURITE.ID }, true);
+}
