+++ stroom-explorer/stroom-explorer-fav-impl-db-jooq/src/main/java/stroom/explorer/fav/impl/db/jooq/Stroom.java
@@ -1,0 +1,54 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package stroom.explorer.fav.impl.db.jooq;
+
+
+import java.util.Arrays;
+import java.util.List;
+
+import org.jooq.Catalog;
+import org.jooq.Table;
+import org.jooq.impl.SchemaImpl;
+
+import stroom.explorer.fav.impl.db.jooq.tables.ExplorerFavourite;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class Stroom extends SchemaImpl {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * The reference instance of <code>stroom</code>
+     */
+    public static final Stroom STROOM = new Stroom();
+
+    /**
+     * The table <code>stroom.explorer_favourite</code>.
+     */
+    public final ExplorerFavourite EXPLORER_FAVOURITE = ExplorerFavourite.EXPLORER_FAVOURITE;
+
+    /**
+     * No further instances allowed
+     */
+    private Stroom() {
+        super("stroom", null);
+    }
+
+
+    @Override
+    public Catalog getCatalog() {
+        return DefaultCatalog.DEFAULT_CATALOG;
+    }
+
+    @Override
+    public final List<Table<?>> getTables() {
+        return Arrays.asList(
+            ExplorerFavourite.EXPLORER_FAVOURITE
+        );
+    }
+}
