+++ stroom-explorer/stroom-explorer-fav-impl-db-jooq/src/main/java/stroom/explorer/fav/impl/db/jooq/Tables.java
@@ -1,0 +1,20 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package stroom.explorer.fav.impl.db.jooq;
+
+
+import stroom.explorer.fav.impl.db.jooq.tables.ExplorerFavourite;
+
+
+/**
+ * Convenience access to all tables in stroom.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class Tables {
+
+    /**
+     * The table <code>stroom.explorer_favourite</code>.
+     */
+    public static final ExplorerFavourite EXPLORER_FAVOURITE = ExplorerFavourite.EXPLORER_FAVOURITE;
+}
