+++ stroom-explorer/stroom-explorer-fav-impl-db-jooq/src/main/java/stroom/explorer/fav/impl/db/jooq/tables/records/ExplorerFavouriteRecord.java
@@ -1,0 +1,255 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package stroom.explorer.fav.impl.db.jooq.tables.records;
+
+
+import org.jooq.Field;
+import org.jooq.Record1;
+import org.jooq.Record5;
+import org.jooq.Row5;
+import org.jooq.impl.UpdatableRecordImpl;
+
+import stroom.explorer.fav.impl.db.jooq.tables.ExplorerFavourite;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class ExplorerFavouriteRecord extends UpdatableRecordImpl<ExplorerFavouriteRecord> implements Record5<Long, String, String, String, Long> {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * Setter for <code>stroom.explorer_favourite.id</code>.
+     */
+    public void setId(Long value) {
+        set(0, value);
+    }
+
+    /**
+     * Getter for <code>stroom.explorer_favourite.id</code>.
+     */
+    public Long getId() {
+        return (Long) get(0);
+    }
+
+    /**
+     * Setter for <code>stroom.explorer_favourite.doc_type</code>.
+     */
+    public void setDocType(String value) {
+        set(1, value);
+    }
+
+    /**
+     * Getter for <code>stroom.explorer_favourite.doc_type</code>.
+     */
+    public String getDocType() {
+        return (String) get(1);
+    }
+
+    /**
+     * Setter for <code>stroom.explorer_favourite.doc_uuid</code>.
+     */
+    public void setDocUuid(String value) {
+        set(2, value);
+    }
+
+    /**
+     * Getter for <code>stroom.explorer_favourite.doc_uuid</code>.
+     */
+    public String getDocUuid() {
+        return (String) get(2);
+    }
+
+    /**
+     * Setter for <code>stroom.explorer_favourite.user_uuid</code>.
+     */
+    public void setUserUuid(String value) {
+        set(3, value);
+    }
+
+    /**
+     * Getter for <code>stroom.explorer_favourite.user_uuid</code>.
+     */
+    public String getUserUuid() {
+        return (String) get(3);
+    }
+
+    /**
+     * Setter for <code>stroom.explorer_favourite.create_time_ms</code>.
+     */
+    public void setCreateTimeMs(Long value) {
+        set(4, value);
+    }
+
+    /**
+     * Getter for <code>stroom.explorer_favourite.create_time_ms</code>.
+     */
+    public Long getCreateTimeMs() {
+        return (Long) get(4);
+    }
+
+    // -------------------------------------------------------------------------
+    // Primary key information
+    // -------------------------------------------------------------------------
+
+    @Override
+    public Record1<Long> key() {
+        return (Record1) super.key();
+    }
+
+    // -------------------------------------------------------------------------
+    // Record5 type implementation
+    // -------------------------------------------------------------------------
+
+    @Override
+    public Row5<Long, String, String, String, Long> fieldsRow() {
+        return (Row5) super.fieldsRow();
+    }
+
+    @Override
+    public Row5<Long, String, String, String, Long> valuesRow() {
+        return (Row5) super.valuesRow();
+    }
+
+    @Override
+    public Field<Long> field1() {
+        return ExplorerFavourite.EXPLORER_FAVOURITE.ID;
+    }
+
+    @Override
+    public Field<String> field2() {
+        return ExplorerFavourite.EXPLORER_FAVOURITE.DOC_TYPE;
+    }
+
+    @Override
+    public Field<String> field3() {
+        return ExplorerFavourite.EXPLORER_FAVOURITE.DOC_UUID;
+    }
+
+    @Override
+    public Field<String> field4() {
+        return ExplorerFavourite.EXPLORER_FAVOURITE.USER_UUID;
+    }
+
+    @Override
+    public Field<Long> field5() {
+        return ExplorerFavourite.EXPLORER_FAVOURITE.CREATE_TIME_MS;
+    }
+
+    @Override
+    public Long component1() {
+        return getId();
+    }
+
+    @Override
+    public String component2() {
+        return getDocType();
+    }
+
+    @Override
+    public String component3() {
+        return getDocUuid();
+    }
+
+    @Override
+    public String component4() {
+        return getUserUuid();
+    }
+
+    @Override
+    public Long component5() {
+        return getCreateTimeMs();
+    }
+
+    @Override
+    public Long value1() {
+        return getId();
+    }
+
+    @Override
+    public String value2() {
+        return getDocType();
+    }
+
+    @Override
+    public String value3() {
+        return getDocUuid();
+    }
+
+    @Override
+    public String value4() {
+        return getUserUuid();
+    }
+
+    @Override
+    public Long value5() {
+        return getCreateTimeMs();
+    }
+
+    @Override
+    public ExplorerFavouriteRecord value1(Long value) {
+        setId(value);
+        return this;
+    }
+
+    @Override
+    public ExplorerFavouriteRecord value2(String value) {
+        setDocType(value);
+        return this;
+    }
+
+    @Override
+    public ExplorerFavouriteRecord value3(String value) {
+        setDocUuid(value);
+        return this;
+    }
+
+    @Override
+    public ExplorerFavouriteRecord value4(String value) {
+        setUserUuid(value);
+        return this;
+    }
+
+    @Override
+    public ExplorerFavouriteRecord value5(Long value) {
+        setCreateTimeMs(value);
+        return this;
+    }
+
+    @Override
+    public ExplorerFavouriteRecord values(Long value1, String value2, String value3, String value4, Long value5) {
+        value1(value1);
+        value2(value2);
+        value3(value3);
+        value4(value4);
+        value5(value5);
+        return this;
+    }
+
+    // -------------------------------------------------------------------------
+    // Constructors
+    // -------------------------------------------------------------------------
+
+    /**
+     * Create a detached ExplorerFavouriteRecord
+     */
+    public ExplorerFavouriteRecord() {
+        super(ExplorerFavourite.EXPLORER_FAVOURITE);
+    }
+
+    /**
+     * Create a detached, initialised ExplorerFavouriteRecord
+     */
+    public ExplorerFavouriteRecord(Long id, String docType, String docUuid, String userUuid, Long createTimeMs) {
+        super(ExplorerFavourite.EXPLORER_FAVOURITE);
+
+        setId(id);
+        setDocType(docType);
+        setDocUuid(docUuid);
+        setUserUuid(userUuid);
+        setCreateTimeMs(createTimeMs);
+    }
+}
