+++ stroom-explorer/stroom-explorer-impl-db/build.gradle
@@ -5,7 +5,7 @@ dependencies {
     implementation project(':stroom-core-shared')
     implementation project(':stroom-db-util')
     implementation project(':stroom-docref')
-    implementation project(':stroom-docstore:stroom-docstore-fav-impl-db-jooq')
+    implementation project(':stroom-explorer:stroom-explorer-fav-impl-db-jooq')
     implementation project(':stroom-explorer:stroom-explorer-impl')
     implementation project(':stroom-explorer:stroom-explorer-impl-db-jooq')
     implementation project(':stroom-security:stroom-security-api')
